base/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.12.3
bleach==4.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.17
django-admin-interface==0.28.9
django-appconf==1.0.6
django-colorfield==0.11.0
django-compressor==4.5.1
django-image-optimizer==1.0.3
django-import-export==4.0.10
django-summernote==0.8.20.0
et-xmlfile==2.0.0
funcy==2.0
html5lib==1.1
idna==3.7
openpyxl==3.1.5
pilkit==3.0
pillow==10.4.0
PyMySQL==1.1.1
python-decouple==3.8
python-resize-image==1.1.20
python-slugify==8.0.4
pytz==2024.2
rcssmin==1.1.2
redis==5.0.1
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
typing-extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
